Split eCFR transforms package and remove shim module
