+++ solution/parsers/ecfr-worker/transforms/__init__.py
@@ -1,0 +1,12 @@
+"""Public transform API for eCFR worker structure processing."""
+
+from .errors import EcfrTransformError
+from .locations import determine_part_depth, extract_sections_and_subparts
+from .normalize import normalize_structure_for_upload
+
+__all__ = [
+    "EcfrTransformError",
+    "determine_part_depth",
+    "extract_sections_and_subparts",
+    "normalize_structure_for_upload",
+]
